Logging in .Framework not .AspNet
