+++ src/Microsoft.Framework.Logging.Testing/Microsoft.Framework.Logging.Testing.kproj
