+++ src/Microsoft.Framework.Logging.Testing/NullLoggerFactory.cs
@@ -1,25 +1,17 @@
 ﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-using Microsoft.Framework.Logging;
 
-namespace Microsoft.AspNet.Testing.Logging
+namespace Microsoft.Framework.Logging.Testing
 {
-    public class TestLoggerFactory : ILoggerFactory
+    public class NullLoggerFactory : ILoggerFactory
     {
-        private readonly TestSink _sink;
-        private readonly bool _enabled;
-
-        public TestLoggerFactory(TestSink sink, bool enabled)
-        {
-            _sink = sink;
-            _enabled = enabled;
-        }
+        public static readonly NullLoggerFactory Instance = new NullLoggerFactory();
 
         public LogLevel MinimumLevel { get; set; } = LogLevel.Verbose;
 
         public ILogger CreateLogger(string name)
         {
-            return new TestLogger(name, _sink, _enabled);
+            return NullLogger.Instance;
         }
 
         public void AddProvider(ILoggerProvider provider)
