+++ src/Microsoft.Framework.Logging.Testing/TestLogger.cs
@@ -2,9 +2,8 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.Framework.Logging;
 
-namespace Microsoft.AspNet.Testing.Logging
+namespace Microsoft.Framework.Logging.Testing
 {
     public class TestLogger : ILogger
     {
