+++ src/Microsoft.Framework.Logging.Testing/WriteContext.cs
@@ -1,7 +1,9 @@
-﻿using System;
-using Microsoft.Framework.Logging;
+﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNet.Testing.Logging
+using System;
+
+namespace Microsoft.Framework.Logging.Testing
 {
     public class WriteContext
     {
