+++ src/Microsoft.Framework.Logging.Testing/project.json
