+++ test/Microsoft.Framework.Logging.Testing.Tests/Microsoft.Framework.Logging.Testing.Tests.kproj
