+++ test/Microsoft.Framework.Logging.Testing.Tests/NullLoggerFactoryTest.cs
@@ -2,10 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.Framework.Logging;
 using Xunit;
 
-namespace Microsoft.AspNet.Testing.Logging
+namespace Microsoft.Framework.Logging.Testing
 {
     public class NullLoggerFactoryTest
     {
