+++ test/Microsoft.Framework.Logging.Testing.Tests/project.json
@@ -1,7 +1,7 @@
 ﻿{
     "version": "1.0.0-*",
     "dependencies": {
-        "Microsoft.AspNet.Testing.Logging": "1.0.0-*",
+        "Microsoft.Framework.Logging.Testing": "1.0.0-*",
         "xunit.runner.kre": "1.0.0-*"
     },
     "commands": {
